+++ tests/testthat/test-read-csv.R
@@ -253,3 +253,8 @@ test_that("read_csv does not duplicate header rows for leading whitespace (747)"
   expect_equal(nrow(x), 1)
   expect_equal(x$foo, 1)
 })
+
+test_that("read_csv handles whitespace between delimiters and quoted fields (668)", {
+  x <- read_csv('x,y\n1, \"hi,there\"\n3,4')
+  expect_equal(x$y, c("hi,there", "4"))
+})